omit ones which are already processed, add try-except block

+++ offline.py
@@ -7,8 +7,14 @@ from feature_extractor import FeatureExtractor
 fe = FeatureExtractor()
 
 for img_path in sorted(glob.glob('static/img/*.jpg')):
-    print(img_path)
-    img = Image.open(img_path)  # PIL image
-    feature = fe.extract(img)
-    feature_path = 'static/feature/' + os.path.splitext(os.path.basename(img_path))[0] + '.pkl'
-    pickle.dump(feature, open(feature_path, 'wb'))
+    pkl_file_path = os.path.join('static/feature/', os.path.splitext(os.path.basename(img_path))[0] + '.pkl')
+    if not os.path.exists(pkl_file_path):
+        try:
+            print(img_path)
+            img = Image.open(img_path)  # PIL image
+            feature = fe.extract(img)
+            feature_path = 'static/feature/' + os.path.splitext(os.path.basename(img_path))[0] + '.pkl'
+            pickle.dump(feature, open(feature_path, 'wb'))
+        except:
+            pass
+
